Remove nova.compute.*API() shims
When we had cellsv1, we had two different sets of
[Host|InstanceAction|ComputeCells]API. Now that cellsv1 is gone, we only
need nova.compute.api.API, .HostAPI, and .InstanceActionAPI. This patch
removes the dynamic loader shims from nova/compute/__init__.py and swaps
out all references to directly access the classes in nova.compute.api.

Note that there are a couple of ways we could have done this. One way
would have been to replace

  from nova import compute

with

  from nova.compute import api

and then change

  self.compute_api = compute.API()

to

  self.compute_api = api.API()

However, the current approach was preferred because
- smaller delta
- the code reads better as compute.API(), which is more important than
  the import being pretty

cleanup for blueprint remove-cells-v1

Change-Id: I84d9692efa3a131d6392dbd1011dfc43e4ac0b53