Remove PlacementAPIConnectFailure handling from AggregateAPI
Change Ibd7aa4f8c4ea787774becece324d9051521c44b6 in Rocky
started mirroring aggregate membership changes to the placement
service but to ease upgrades (because before that nova-api didn't
rely on placement) the API code would handle failures to connect
to the placement service. A TODO was added to remove that handling
in Stein, so it's time to remove it.

Note that with the removal of that connection error handling, the
API response in both add host to aggregate and remove host from
aggregate APIs will be a 500. An error in add_host_to_aggregate
should be remedied by running the "nova-manage placement sync_aggregates"
command but a failure in remove_host_from_aggregate will not be [1]. As
such the following in the remove_host_from_aggregate operation is changed:

* ResourceProviderNotFound is still ignored since if we get that the
  provider likely no longer exists and is not in an aggregate anyway.
* ResourceProviderUpdateConflict is handled in the route handler to return
  a 409 rather than 500 error. Anything else raised from
  SchedulerReportClient.aggregate_remove_host results in a 500 error.
* The placement work is done first because if that fails the admin can
  try again from the compute API. Otherwise if we do the placement
  cleanup after Aggregate.delete_host is called in nova, the admin cannot
  retry the remove_host action because it will fail with
  AggregateHostNotFound.

[1] https://opendev.org/openstack/nova/src/tag/20.0.0/nova/cmd/manage.py#L2533

Change-Id: I4dfbed3e41c4952efc2900a063cb22753539d6cb