Validate requested host/node during servers create
This change adds a method to validate a requested host and/or
hypervisor_hostname(node) during servers create. We are simply
checking for existence of the resource, so the host is validated
for the HostMapping and the node is validated by checking for
the related resource provider in the placement service.

Part of Blueprint: add-host-and-hypervisor-hostname-flag-to-create-server

Change-Id: I009761162817c6ed1035d2208605268dc82d430f