Skip novnc tests in multi-cell job until bug 1830417 is fixed
NoVNCConsoleTestJSON.test_novnc is intermittently failing
in the multi-cell job when trying to get a vnc console for
an instance in cell2. It looks like requests aren't getting
routed to the novnc-cell2 service properly, so let's skip
novnc tests until the bug is fixed since we want to stabilize
this job to make it voting.

Change-Id: I74a4b392ccec66a7a80a117d1cdf83a418c7dcdc
Related-Bug: #1830417