Reduce logging of host hypervisor capabilities to DEBUG level
While the results of invoking libvirt's getCapabilities() API call get
deliberately logged at INFO level[0], doing the same for
getDomainCapabilities() is one bridge too far, because we call it once
per (architecture, default_machine_type) tuple.

So heed the suggestions of Stephen and Eric[1] and reduce it to DEBUG level.

[0] See Ica0dccbfbff89829a7e631adaaf2cde13c449773 at
    https://review.opendev.org/#/c/158334/
[1] https://review.opendev.org/#/c/655268/3/nova/virt/libvirt/host.py@777

Change-Id: Idb4c4380c32f78d639f41cb1b5204578e1a19cd3