extract baselineCPU API call from _get_cpu_traits()
_get_cpu_traits() in the libvirt driver was quite complicated, dealing
with multiple modes of the guest CPU model config, and also invoking
libvirt's baselineCPU API.  This made it hard to see at a glance the
flow through the method, and which variables were used where.

So extract the baselineCPU API call into a separate method, resulting
in clearer variable scoping.  Also document the API call and its
inputs and outputs, based on the official API docs:

https://libvirt.org/html/libvirt-libvirt-host.html#virConnectBaselineCPU

Change-Id: I6f42da5e2fd2e65d31b8e9191e3dff98b011d0bf