Workaround missing RequestSpec.instance_group.uuid
It's clear that we could have a RequestSpec.instance_group
without a uuid field if the InstanceGroup is set from the
_populate_group_info method which should only be used for
legacy translation of request specs using legacy filter
properties dicts.

To workaround the issue, we look for the group scheduler hint
to get the group uuid before loading it from the DB.

The related functional regression recreate test is updated
to show this solves the issue.

Change-Id: I20981c987549eec40ad9762e74b0db16e54f4e63
Closes-Bug: #1830747