Add functional test for anti-affinity cross-cell migration
This adds a functional test to ensure that the scheduler
properly restricts anti-affinity group members to different
hosts regardless of which cell those hosts are in. The comments
in the setup_instance_group scheduler utility method assume
that all moves are within the same cell and as such instance
group hosts will be within the same cell, which is not true in
the case of an anti-affinity group cross-cell move. The test
added here would actually fail were it not for change
I4b67ec9dd4ce846a704d0f75ad64c41e693de0fb in Rocky to the
ServerGroupAntiAffinityFilter such that the filter does not
look at the request_spec.instance_group.hosts but only the
members. A clarifying comment is added to the scheduler utils
code in case we ever need to fix that.

Part of blueprint cross-cell-resize

Change-Id: I29f0dfdc52fda6d0b7c378801ee40ec7ad8fdf40