xenapi: log quality warning in init_host
The xenapi driver has not had working third party CI
for several months. As a result, this change logs a
warning on startup of the service in the init_host
method indicating the lack of testing and the possibility
that the driver may be deprecated in the future if the
situation does not change.

This was discussed at the 2019-05-30 nova meeting [1] and
also at the Train PTG.

[1] http://eavesdrop.openstack.org/meetings/nova/2019/nova.2019-05-30-14.00.log.html

Change-Id: I7f8eb7d5c5a9b1cb0a8d5e607d719b49a22675d3