Change the default of notification_format to unversioned
The default config `both` means that both the legacy and the versioned
notifications are emitted. This was selected as default in the past when
we thought that this will help the adoption of the versioned interface
while we worked on to make that new interface in feature parity with the
legacy. Even though the versioned notification interface is in feature
parity with the legacy interface since Stein the projects consuming nova
notifications do not have the resources to switch to the new interface.

In the other hand having `both` as a default in an environtment where
only the legacy notifications are consumed causes performance issues in
the message bus hence the bug #1805659.

The original plan was that we set the default to `versioned` when the
interface reaches feature parity but as major consumers are not ready
to switch we cannot do that.

So the only option left is to set the default to `unversioned`.

Related devstack patch: https://review.opendev.org/#/c/662849/

Closes-Bug: #1805659

Change-Id: I72faa356afffb7a079a9ce86fed1b463773a0507