Unplug VIFs as part of cleanup of networks
If an instance fails to build, which is possible for a variety of
reasons, we may end up in a situation where we have remnants of a
plugged VIF (typically files) left on the host. This is because we
cleanup from the neutron perspective but don't attempt to unplug the
VIF, a call which may have many side-effects depending on the VIF
driver. Resolve this by always attempting to unplug VIFs as part of the
network cleanup.

A now invalid note is also removed and a unit test corrected.

Conflicts:
	nova/tests/unit/compute/test_compute_mgr.py

NOTE(stephenfin): Conflicts are due to the absence of change
Ifa9c5c468400261a5e1f66b72c575845173a4f8f ("nova-net: Remove final
references to nova-network") which we don't want to backport here. In
addition, we need to modify a mock to reflect the absence of change
I329f0fd589a4b2e0426485f09f6782f94275cc07 ("nova-net: Remove layer of
indirection in 'nova.network'").

Closes-Bug: #1831771
Related-Bug: #1830081
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
Change-Id: Ibdbde4ed460a99b0cbe0d6b76e0e5b3c0650f9d9
(cherry picked from commit b3e14931d6aac6ee5776ce1e6974c75a5a6b1823)