Delete unused get_all_host_states method
The CachingScheduler was the last thing to use the
HostManager.get_all_host_states method and the
CachingScheduler was deleted in Stein [1].

This removes the unused get_all_host_states method.
The majority of the change is in the test code that
was using get_all_host_states as an entry point to
test the _get_host_states method, so rather than
outright delete the tests and potentially drop coverage
on _get_host_states, those tests are refactored a bit
to call _get_host_states directly.

[1] I1832da2190be5ef2b04953938860a56a43e8cddf

Change-Id: Ie19f1be2d60d4d20fac99508d1683ab3f266bfba