Convert HostMapping.cells to a dict
This changes the HostMapping.cells list to a dict,
keyed by cell uuid, to each CellMapping object.
This is going to be used in a later change [1] where
we will create a cached mapping of host to cell uuid
so we can go quickly from compute node host value to
get the CellMapping without having to lookup the
HostMapping again.

[1] Ic6b1edfad2e384eb32c6942edc522ee301123cbc

Change-Id: I75e5bf2c45688f68eeca27404641f4c438b023c6
Related-Bug: #1737465