Skip recomputing baremetal quota usage
We're using a single function to compute the usage of multiple quota
reources. There's already a check in place that prohibits recomputing
that data, if the resource's name is already in the data. When computing
usage data, we use the instances belonging to a user/project. If that
user/project does not have any baremetal instances of a flavor, we don't
add the flavor's resource to the usage data. This means, the check is
unable to skip recomputing the data.

If "instances" is already in the usage data, our shared function must
have been called already. Therefore, if we encounter a resource having a
name starting with "instances_" - which should be only true for
resources we created for baremetal flavors - we can skip the
recomputation even if the resource's name is not in the usage data.

(cherry picked from commit 3a820fdab91d1c4e64efaea42ed0ec8f75d6fa6c)