Make get_provider_by_name public and remove safe_connect
This makes _get_provider_by_name public to expose access for
callers outside of SchedulerReportClient. The @safe_connect
decorator is removed and the get() is wrapped in a try/except
to handle KSA ClientException errors and raise
PlacementAPIConnectFailure.

The two existing calls for adding/removing providers to/from
aggregates are updated to use the public method so they don't
need to individually do the safe_connect handling.

Change-Id: I504c374d3863a2a956d5c0156a43be2d2a2bc712