Skip test_check_doubled_words hacking check UT
The referenced bug has reared its ugly head in python 2.7.15 in (at
least) releaselevel='final', which is being used in the gate, as well as
3.6.7+. Disable the test entirely until we can fix or work around the
bug in the tokenizer.

The hacking check itself is still in place. This is just disabling the
unit test that tries to test the hacking check.

NOTE: Noticed that the hacking check is operating on physical_line,
which means it'll catch things like:

 "This is the the error"

but not:

 ("This is the "
  "the error")

...which is probably way more likely.

Change-Id: I8826c3fb89690805baae6b9b7b48985abb8d62d3
Related-Bug: #1804062