Remove file-backed memory live migration compat check
This removes a compat check for pre-Rocky computes added
in I676291ec0faa1dea0bd5050ef8e3426d171de4c6 for file-backed
memory live migration support. Now that we're in Train we can
assume that we won't be live migrating with computes older
than Rocky and can remove the check.

Change-Id: Iefd7a60139043929aee63a3660fabdded1622029