tests: Use consistent URL regex substitution
For some reason, we were using a really convoluted regex that allowed
URLs of formats like 'https:\\foo' (i.e. backslashes instead of forward
slashes). I'm not really sure if this is even valid but it's certainly
not necessary. Use a far simpler substitution and DRY the test cases up
a little.

Change-Id: I634a6de795893d995c4e7c3707f73061c7aa5401
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>