Grab fresh power state info from the driver
In drivers that use a cache to store the node info (presently only
ironic since [1]), the "_get_power_state" function called during
instance actions like start or stop grabs the information from the node
cache and saves it in the nova database instead of getting fresh
information from the driver. This leads to inconsistency between
the vm_state and power_state for an instance in the nova database
(which remains until a power_sync happens between nova and ironic).
This can be confusing for a user when doing "nova list" where the
power_state might still be shutdown when the vm_state has already
become active. On a default environment this inconsistency lasts
for about ten minutes which is the default value for the
sync_power_state_interval interval.

This patch changes the "use_cache" to False in the compute manager
when triggering an action on an instance like start/stop/reboot.

[1] I0069cbc327d952d42dbb8fe54949faab89995a7e

Change-Id: I8bca5d84c37d02331d2f9968a674f3398c1a8f5b
Closes-Bug: #1832720
(cherry picked from commit 5b1c9dd05ccc0488e4d47bfc2442e46adb82b765)