Revert "Revert resize: wait for events according to hybrid plug"
This reverts commit 19f9b37721d9bc13bc1ed35a4f368b1d21b10a5b.

Commit 19f9b37721d9bc13bc1ed35a4f368b1d21b10a5b introduced
a regression and caused errors on Neutron CI jobs which run
on a single node with iptables_hybrid firewall driver.
This is because the nova change made the resize revert flow
wait for neutron-vif-plugged events in the ComputeManager
when the port's host binding changes for iptables_hybrid
ports but for those types of ports, Neutron does not send
the event if the host does not change - which is the case
for a same-host resize being reverted.

Change-Id: I77b3639435c671bacca9cdfc1aa203e44a2fb042
Closes-Bug: #1833902