xenapi: implement update_provider_tree
This replaces get_inventory with update_provider_tree
in the xenapi driver. Since the driver is not reporting
traits there is nothing to do with traits in upt. The
VCPU, MEMORY_MB, DISK_GB and VGPU inventory is reported
the same as before, i.e. no support for nested providers
or reshaping allocations. The only additional change is
the need to provide the allocation_ratio and reserved
inventory values for VCPU, MEMORY_MB and DISK_GB since
ResourceTracker._normalize_inventory_from_cn_obj is not
called when the driver implements update_provider_tree.

Change-Id: Ie7bc0b3d3c3d5b1cdea124f6415aa0e87ac8b757