Deprecate non-update_provider_tree compat code
With all in-tree virt drivers now implementing the
update_provider_tree interface, we can deprecate the
compatibility code in the ResourceTracker. This change
simply logs a warning if the driver does not implement
the upt interface and sets the timer for removal in the
U release at the earliest.

The resource tracker unit tests will need to be cleaned
up but that can happen in a separate change so it does
not slow down this deprecation.

Change-Id: I1eae47bce08f6292d38e893a2122289bcd6f4b58